Make sqlcipher password optional

+++ example/src/tests/hooks.spec.ts
@@ -2,7 +2,7 @@ import Chance from 'chance';
 
 import {type DB, open, isLibsql} from '@op-engineering/op-sqlite';
 import chai from 'chai';
-import {describe, it, beforeEach, afterAll} from './MochaRNAdapter';
+import {describe, it, beforeEach, afterEach} from './MochaRNAdapter';
 import {sleep} from './utils';
 
 const expect = chai.expect;
@@ -16,13 +16,12 @@ let db: DB;
 
 export function registerHooksTests() {
   describe('Hooks', () => {
+    if (isLibsql()) {
+      return;
+    }
+
     beforeEach(async () => {
       try {
-        if (db) {
-          db.close();
-          db.delete();
-        }
-
         db = open(DB_CONFIG);
 
         await db.execute('DROP TABLE IF EXISTS User;');
@@ -34,18 +33,11 @@ export function registerHooksTests() {
       }
     });
 
-    afterAll(() => {
+    afterEach(() => {
       if (db) {
-        db.close();
         db.delete();
-        // @ts-ignore
-        db = null;
       }
     });
-    // libsql does not support hooks
-    if (isLibsql()) {
-      return;
-    }
 
     it('update hook', async () => {
       let promiseResolve: any;
@@ -57,18 +49,9 @@ export function registerHooksTests() {
       }>(resolve => {
         promiseResolve = resolve;
       });
-      let db = open({
-        name: 'updateHookDb.sqlite',
-        encryptionKey: 'blah',
-      });
-
-      await db.execute('DROP TABLE IF EXISTS User;');
-
-      await db.execute(
-        'CREATE TABLE User ( id INT PRIMARY KEY, name TEXT NOT NULL, age INT, networth REAL) STRICT;',
-      );
 
       db.updateHook(data => {
+        console.log('UPDATE HOOK CALLED');
         promiseResolve(data);
       });
 
@@ -76,17 +59,21 @@ export function registerHooksTests() {
       const name = chance.name();
       const age = chance.integer();
       const networth = chance.floating();
-      await db.execute(
-        'INSERT INTO "User" (id, name, age, networth) VALUES(?, ?, ?, ?)',
-        [id, name, age, networth],
-      );
+      await db.transaction(async tx => {
+        await tx.execute(
+          'INSERT INTO "User" (id, name, age, networth) VALUES(?, ?, ?, ?)',
+          [id, name, age, networth],
+        );
+      });
+
+      console.log('AWAITING UPDATE HOOK PROMISE');
 
       const data = await promise;
 
       expect(data.operation).to.equal('INSERT');
       expect(data.rowId).to.equal(1);
 
-      db.close();
+      db.updateHook(null);
     });
 
     it('remove update hook', async () => {
@@ -103,10 +90,12 @@ export function registerHooksTests() {
       const name = chance.name();
       const age = chance.integer();
       const networth = chance.floating();
+      // await db.transaction(async tx => {
       await db.execute(
         'INSERT INTO "User" (id, name, age, networth) VALUES(?, ?, ?, ?)',
         [id, name, age, networth],
       );
+      // });
 
       db.updateHook(null);
 
@@ -118,7 +107,6 @@ export function registerHooksTests() {
       await sleep(0);
 
       expect(hookRes.length).to.equal(1);
-      db.close();
     });
 
     it('commit hook', async () => {
@@ -143,6 +131,7 @@ export function registerHooksTests() {
       });
 
       await promise;
+      db.commitHook(null);
     });
 
     it('remove commit hook', async () => {
